+++ CHANGES.md
@@ -5,6 +5,8 @@
   * Added Request#withFile(String, byte[]), Request#withFile(String, ByteBuffer)
   * Added Request#withEntity(HttpEntity) and Request#withContent(String, String)
   * Added PutResource example
+  * Added setting to change the default content type for requests
+    (CloudAPI#setDefaultContentType(String))
 
 ## 1.0.0 2011-05-19